requirements.txt
asgiref==3.5.2
autopep8==1.7.0
backports.zoneinfo==0.2.1
certifi==2022.9.24
charset-normalizer==2.1.1
dhis2.py==2.3.0
Django==4.1
django-colorfield==0.7.2
django-cors-headers==3.13.0
django-extensions==3.2.1
django-filter==22.1
djangorestframework==3.13.1
djangorestframework-simplejwt==5.2.0
gunicorn==20.1.0
idna==3.4
isort==5.10.1
logzero==1.7.0
numpy==1.23.5
pandas==1.5.2
Pillow==9.2.0
psycopg2-binary==2.9.5
pycodestyle==2.9.1
Pygments==2.13.0
PyJWT==2.4.0
python-dateutil==2.8.2
pytz==2022.2.1
requests==2.28.1
six==1.16.0
sqlparse==0.4.2
support-developer==1.0.5
toml==0.10.2
transpose-dict==1.1.3
urllib3==1.26.12
